+++ src/Kmd.Logic.Identity.Authorization/LogicTokenProviderOptions.cs
@@ -13,15 +13,12 @@ namespace Kmd.Logic.Identity.Authorization
         /// <remarks>
         /// This option should not be overridden except for testing purposes.
         /// </remarks>
-        public Uri AuthorizationTokenIssuer { get; set; } = new Uri("https://login.microsoftonline.com/logicidentityprod.onmicrosoft.com/oauth2/v2.0/token");
+        public Uri AuthorizationTokenIssuer { get; set; } = new Uri("https://identity-api.kmdlogic.io/clientcredentials/token");
 
         /// <summary>
         /// Gets or sets the Logic Identity authorization scope.
         /// </summary>
-        /// <remarks>
-        /// This option should not be overridden except for testing purposes.
-        /// </remarks>
-        public string AuthorizationScope { get; set; } = "https://logicidentityprod.onmicrosoft.com/bb159109-0ccd-4b08-8d0d-80370cedda84/.default";
+        public string AuthorizationScope { get; set; }
 
         /// <summary>
         /// Gets or sets the client credentials identifier.