+++ CHANGES.md
@@ -1,2 +1,8 @@
+2.0.0
+* Removed the default AuthorizationScope [BREAKING CHANGE]
+* To maintain backward compatibility, added DefaultAuthorizationScope to allow referencing libraries to provide an AuthorizationScope
+* Added the sample application
+* Moved build environment to dotnet core 3.1
+
 1.0.0
 * Initial version